pyproject.toml
[tool.poetry]
name = "atcodercli"
version = "0.5.1"
description = "A lightweight, fast and beautiful command line interface for Atcoder.jp"
authors = ["Rickyxrc <rickyxrc@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
rich = "^13.7.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"
babel = "^2.14.0"
black = "^23.12.1"
isort = "^5.13.2"
pre-commit = "^3.6.0"

[tool.poetry.build]
generate-setup-file = false
script = "scripts/build-mo.py"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
